+++ src/main/java/kitchenpos/order/dao/OrderLineItemRepository.java
@@ -1,0 +1,9 @@
+package kitchenpos.order.dao;
+
+import kitchenpos.order.domain.OrderLineItem;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface OrderLineItemRepository extends JpaRepository<OrderLineItem, Long> {
+}
