+++ src/main/java/kitchenpos/order/domain/OrderLineItem.java
@@ -1,0 +1,57 @@
+package kitchenpos.order.domain;
+
+import kitchenpos.common.Quantity;
+import kitchenpos.menu.domain.Menu;
+
+import javax.persistence.*;
+
+@Entity
+public class OrderLineItem {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long seq;
+
+    @ManyToOne
+    @JoinColumn(name = "order_id")
+    private Order order;
+
+    @ManyToOne
+    @JoinColumn(name = "menu_id")
+    private Menu menu;
+
+    @Embedded
+    private Quantity quantity;
+
+    protected OrderLineItem() {
+    }
+
+    public OrderLineItem(Menu menu, long quantity) {
+       this(null, menu, quantity);
+    }
+
+    public OrderLineItem(Long seq, Menu menu, long quantity) {
+        this.seq = seq;
+        this.menu = menu;
+        this.quantity = new Quantity(quantity);
+    }
+
+    public void updateOrder(Order order) {
+        this.order = order;
+    }
+
+    public Long getSeq() {
+        return seq;
+    }
+
+    public Long getOrderId() {
+        return order.getId();
+    }
+
+    public Long getMenuId() {
+        return menu.getId();
+    }
+
+    public long getQuantity() {
+        return quantity.value();
+    }
+}
